Fix broken test in Windows
Due to the nature of Windows using backslash instead of regular slash to
separate the directories, the test would fail because it was using a
hardcoded output. To address that was changed the method to evaluate the
output to a more flexible one, that way the backslash would not be a
problem.