Change the git pre-commit actions to *actually* use stable
This is in response to #6715: e8bb8dffa1b1af05b7be184c820f5c72a2b8c1da
The pre-commit actions have a higher MSRV than the overall project (1.70.0),
so if 1.70.0 is the configured default, this lead to problems (#6730).
This commit fixes the situation by specifying 'stable', whatever that means on the user's system.