build: restore uutils in GNU test files
Update the `util/build-gnu.sh` script to restore the use of uutils
programs that had been temporarily replaced with GNU programs in test
files where the program to replace was not a primary concern. For
example, before this commit we were forcing the use of GNU `stat`
instead of uutils `stat` in the test file
`tests/touch/60-seconds.sh`. After this commit, we use the uutils
`stat` since it is working well enough for this test case.