.github/workflows/CICD.yml: Do not update gcc
The issue referenced has been long fixed, and for reasons not
totally clear to me, blake3 fails after the GCC update (I tried
to dig into this, but couldn't really figure out if this is really
a problem with GCC 15, or with the version provided by MSYS, or
some other side effect of the exact sequence in CI).

Since blake3 CI doesn't do that gcc update (it uses the default gcc ~12
in the github windows image), let's _also_ not do that, and if there's a
real problem with gcc 15+, that'll presumably fail their CI as well.

Fixes #7977.