deny.toml: remove obsolete skips related to rustix
The holdup on crossterm was bumped in d5258d280, bringing rustix to 0.38.44. Meanwhile while
linux-raw-sys is still at 0.4.15 in the lock file, it is unclear to my why it got added in 2c0f97fc3
in the first place as it does not seem to trigger any complaints from `cargo deny`.