tail: show end of device
gnu coreutils will show end of device for block devices.
In current implementation this is skipped as a block devices
are recognized as untailable.

The change ensures that we first try to output the end of any
file and only then tell the observer that the file is not
observable.