lib/mods/locale: fix locale loading
I'm not sure exactly why this is happening, but when I install
coreutils with `make install PREFIX=some-prefix`, before this patch
coreutils would look in the wrong place for locales - it would look in
e.g. "bin/cut/en-US.ftl" instead of in share/locales/cut/en-US.ftl.

This patch fixes the issue, add a test as well.

Fixes #8115.