mknod: set cli specified file mode
libc's mknod function applies the process' umask in the usual way:
the permissions of the created node are `mode & ~umask`.

the umask of uu-mknod is currently only set to 0 (= no mask)
when the supplied mode is not 0o666:
```rust
let set_umask_0 = config.mode & MODE_RW_UGO != MODE_RW_UGO;
```
because `MODE_RW_UGO == 0o666` is also the default `--mode` argument value.

but when `--mode 0666` is given explicitly, the permissions should be set
exactly to the requested bits, and the umask should not be applied.

this patch fixes the issue by tracking if a custom mode was supplied by argument.

interestingly, GNU coreutils issue a call to `fchmodat2` after `mknod` was called,
even though they also set the umask to 0 if a custom mode was specified via `--mode`.
this patch doesn't add that, but it may be necessary for other cases in the future.

fixes #8342