tr: fix high memory use, possible heap exhaustion
Read the input into a statically sized buffer - 8192, matching
GNU - instead of reading until the end of the line, as reading
until the end of the line in a file with no end of line would
result in reading the entire file into memory.

Confusingly, GNU tr seems to write the 8192 byte in two chunks
of 1024 and 7168 byte, but I can't figure out why it would do
that; I don't see any line buffering in GNU tr.

Bug-Ubuntu: https://launchpad.net/bugs/2119520