tr: use an unbuffered stdout
Our stdin that we transform already is buffered (using 8192 byte
buffers in the previous commit), so avoid buffering our output
needlessly.

This effectively changes the code to write complete lines
immediately, for example, in

`( echo a; sleep 1 ) | tr a b`

we receive

    read(0, "a\n", 8192)                    = 2
    write(1, "b\n", 2)                      = 2
    read(0, "", 8192)                       = 0

instead of

    read(0, "a\n", 8192)                    = 2
    read(0, "", 8192)                       = 0
    write(1, "b\n", 2)                      = 2

which matches the GNU coreutils behavior.