cp: force: open dest. to check for required delete
So far cp only checked if the metadata of the destination file indicates
read-only permissions. In order to handle the case where the destination
file is in use, we try to open the destination file for writing. Also
other cases which prevent overwriting can be detected that way.

For the test case we copy the coreutils binary to a test directory. Then
use the copy to try to copy a file over itself. This should fail in
non-force mode, but succeed in force mode.