base_common: re-using input from has_padding
Base64: the function has_padding reads the file and then discards it.
The functions fast_encode and fast_decode re-read the file, providing
significant delay in larger files.
This commit  also reduces the amount of computation done inside the
fast_decode function.
In the next commit there will be the fix for the fast_decode function

Signed-off-by: Andrea Calabrese <andrea.calabrese@amarulasolutions.com>