tests.rs: Read executable binary location from environment at runtime
This commit changes how the `CARGO_BIN_EXE_coreutils` environment variable
is read. Previously, it was read using the `env!` macro, which returns the
value of the variable as exposed by cargo during compilation.

The problem with that is that it renders the process of testing for cross
compiled targets, similar to what build systems like Yocto do, more
difficult.

The solution to this is to read said variable at runtime and if it is not
defined, to default to the version offered by cargo. This way ensures that
first of all, current functionality remains unchanged and that the code can
be more easily cross compiled and tested in environments outside of cargo.

Also, this commit closes #9246 .

Signed-off-by: Andreas Stergiopoulos <andreas.stergiopoulos@smile.fr>