Do not use `.display()`/`.to_string_lossy()` unnecessarily
This comes up a lot for quoted strings in messages: OS strings can be
quoted directly and this prevents information loss.

This commit removes ~60% of the calls to these methods (modulo
tests). Some of the remaining calls are benign, for example because
they convert solely to check for the presence of ASCII
characters. Others are nontrivial to improve.