comm: fix comparison when reading from pipes
Use case is that two files are piped into comm, i.e. in bash
   comm <(cat file1) <(cat file2)

Before the fix, comm reads from the pipes twice. Once in "fn comm" and
once in "fn are_files_identical". As such, part of the data is skipped
in comparison which leads to wrong output.

This is fixed by skipping the file comparison in case one of the files
is not a regular file.